move inter pod affinity predicate logic to its Filter plugin
