+++ pkg/scheduler/algorithm/predicates/metadata_test.go
@@ -27,248 +27,6 @@ import (
 	st "k8s.io/kubernetes/pkg/scheduler/testing"
 )
 
-func TestPodAffinityMetadata_Clone(t *testing.T) {
-	source := &PodAffinityMetadata{
-		topologyToMatchedExistingAntiAffinityTerms: topologyToMatchedTermCount{
-			{key: "name", value: "machine1"}: 1,
-			{key: "name", value: "machine2"}: 1,
-		},
-		topologyToMatchedAffinityTerms: topologyToMatchedTermCount{
-			{key: "name", value: "nodeA"}: 1,
-			{key: "name", value: "nodeC"}: 2,
-		},
-		topologyToMatchedAntiAffinityTerms: topologyToMatchedTermCount{
-			{key: "name", value: "nodeN"}: 3,
-			{key: "name", value: "nodeM"}: 1,
-		},
-	}
-
-	clone := source.Clone()
-	if clone == source {
-		t.Errorf("Clone returned the exact same object!")
-	}
-	if !reflect.DeepEqual(clone, source) {
-		t.Errorf("Copy is not equal to source!")
-	}
-}
-
-// TestGetTPMapMatchingIncomingAffinityAntiAffinity tests against method getTPMapMatchingIncomingAffinityAntiAffinity
-// on Anti Affinity cases
-func TestGetTPMapMatchingIncomingAffinityAntiAffinity(t *testing.T) {
-	newPodAffinityTerms := func(keys ...string) []v1.PodAffinityTerm {
-		var terms []v1.PodAffinityTerm
-		for _, key := range keys {
-			terms = append(terms, v1.PodAffinityTerm{
-				LabelSelector: &metav1.LabelSelector{
-					MatchExpressions: []metav1.LabelSelectorRequirement{
-						{
-							Key:      key,
-							Operator: metav1.LabelSelectorOpExists,
-						},
-					},
-				},
-				TopologyKey: "hostname",
-			})
-		}
-		return terms
-	}
-	newPod := func(labels ...string) *v1.Pod {
-		labelMap := make(map[string]string)
-		for _, l := range labels {
-			labelMap[l] = ""
-		}
-		return &v1.Pod{
-			ObjectMeta: metav1.ObjectMeta{Name: "normal", Labels: labelMap},
-			Spec:       v1.PodSpec{NodeName: "nodeA"},
-		}
-	}
-	normalPodA := newPod("aaa")
-	normalPodB := newPod("bbb")
-	normalPodAB := newPod("aaa", "bbb")
-	nodeA := &v1.Node{ObjectMeta: metav1.ObjectMeta{Name: "nodeA", Labels: map[string]string{"hostname": "nodeA"}}}
-
-	tests := []struct {
-		name                    string
-		existingPods            []*v1.Pod
-		nodes                   []*v1.Node
-		pod                     *v1.Pod
-		wantAffinityPodsMap     topologyToMatchedTermCount
-		wantAntiAffinityPodsMap topologyToMatchedTermCount
-		wantErr                 bool
-	}{
-		{
-			name:  "nil test",
-			nodes: []*v1.Node{nodeA},
-			pod: &v1.Pod{
-				ObjectMeta: metav1.ObjectMeta{Name: "aaa-normal"},
-			},
-			wantAffinityPodsMap:     make(topologyToMatchedTermCount),
-			wantAntiAffinityPodsMap: make(topologyToMatchedTermCount),
-		},
-		{
-			name:         "incoming pod without affinity/anti-affinity causes a no-op",
-			existingPods: []*v1.Pod{normalPodA},
-			nodes:        []*v1.Node{nodeA},
-			pod: &v1.Pod{
-				ObjectMeta: metav1.ObjectMeta{Name: "aaa-normal"},
-			},
-			wantAffinityPodsMap:     make(topologyToMatchedTermCount),
-			wantAntiAffinityPodsMap: make(topologyToMatchedTermCount),
-		},
-		{
-			name:         "no pod has label that violates incoming pod's affinity and anti-affinity",
-			existingPods: []*v1.Pod{normalPodB},
-			nodes:        []*v1.Node{nodeA},
-			pod: &v1.Pod{
-				ObjectMeta: metav1.ObjectMeta{Name: "aaa-anti"},
-				Spec: v1.PodSpec{
-					Affinity: &v1.Affinity{
-						PodAffinity: &v1.PodAffinity{
-							RequiredDuringSchedulingIgnoredDuringExecution: newPodAffinityTerms("aaa"),
-						},
-						PodAntiAffinity: &v1.PodAntiAffinity{
-							RequiredDuringSchedulingIgnoredDuringExecution: newPodAffinityTerms("aaa"),
-						},
-					},
-				},
-			},
-			wantAffinityPodsMap:     make(topologyToMatchedTermCount),
-			wantAntiAffinityPodsMap: make(topologyToMatchedTermCount),
-		},
-		{
-			name:         "existing pod matches incoming pod's affinity and anti-affinity - single term case",
-			existingPods: []*v1.Pod{normalPodA},
-			nodes:        []*v1.Node{nodeA},
-			pod: &v1.Pod{
-				ObjectMeta: metav1.ObjectMeta{Name: "affi-antiaffi"},
-				Spec: v1.PodSpec{
-					Affinity: &v1.Affinity{
-						PodAffinity: &v1.PodAffinity{
-							RequiredDuringSchedulingIgnoredDuringExecution: newPodAffinityTerms("aaa"),
-						},
-						PodAntiAffinity: &v1.PodAntiAffinity{
-							RequiredDuringSchedulingIgnoredDuringExecution: newPodAffinityTerms("aaa"),
-						},
-					},
-				},
-			},
-			wantAffinityPodsMap: topologyToMatchedTermCount{
-				{key: "hostname", value: "nodeA"}: 1,
-			},
-			wantAntiAffinityPodsMap: topologyToMatchedTermCount{
-				{key: "hostname", value: "nodeA"}: 1,
-			},
-		},
-		{
-			name:         "existing pod matches incoming pod's affinity and anti-affinity - multiple terms case",
-			existingPods: []*v1.Pod{normalPodAB},
-			nodes:        []*v1.Node{nodeA},
-			pod: &v1.Pod{
-				ObjectMeta: metav1.ObjectMeta{Name: "affi-antiaffi"},
-				Spec: v1.PodSpec{
-					Affinity: &v1.Affinity{
-						PodAffinity: &v1.PodAffinity{
-							RequiredDuringSchedulingIgnoredDuringExecution: newPodAffinityTerms("aaa", "bbb"),
-						},
-						PodAntiAffinity: &v1.PodAntiAffinity{
-							RequiredDuringSchedulingIgnoredDuringExecution: newPodAffinityTerms("aaa"),
-						},
-					},
-				},
-			},
-			wantAffinityPodsMap: topologyToMatchedTermCount{
-				{key: "hostname", value: "nodeA"}: 2, // 2 one for each term.
-			},
-			wantAntiAffinityPodsMap: topologyToMatchedTermCount{
-				{key: "hostname", value: "nodeA"}: 1,
-			},
-		},
-		{
-			name:         "existing pod not match incoming pod's affinity but matches anti-affinity",
-			existingPods: []*v1.Pod{normalPodA},
-			nodes:        []*v1.Node{nodeA},
-			pod: &v1.Pod{
-				ObjectMeta: metav1.ObjectMeta{Name: "affi-antiaffi"},
-				Spec: v1.PodSpec{
-					Affinity: &v1.Affinity{
-						PodAffinity: &v1.PodAffinity{
-							RequiredDuringSchedulingIgnoredDuringExecution: newPodAffinityTerms("aaa", "bbb"),
-						},
-						PodAntiAffinity: &v1.PodAntiAffinity{
-							RequiredDuringSchedulingIgnoredDuringExecution: newPodAffinityTerms("aaa", "bbb"),
-						},
-					},
-				},
-			},
-			wantAffinityPodsMap: make(topologyToMatchedTermCount),
-			wantAntiAffinityPodsMap: topologyToMatchedTermCount{
-				{key: "hostname", value: "nodeA"}: 1,
-			},
-		},
-		{
-			name:         "incoming pod's anti-affinity has more than one term - existing pod violates partial term - case 1",
-			existingPods: []*v1.Pod{normalPodAB},
-			nodes:        []*v1.Node{nodeA},
-			pod: &v1.Pod{
-				ObjectMeta: metav1.ObjectMeta{Name: "anaffi-antiaffiti"},
-				Spec: v1.PodSpec{
-					Affinity: &v1.Affinity{
-						PodAffinity: &v1.PodAffinity{
-							RequiredDuringSchedulingIgnoredDuringExecution: newPodAffinityTerms("aaa", "ccc"),
-						},
-						PodAntiAffinity: &v1.PodAntiAffinity{
-							RequiredDuringSchedulingIgnoredDuringExecution: newPodAffinityTerms("aaa", "ccc"),
-						},
-					},
-				},
-			},
-			wantAffinityPodsMap: make(topologyToMatchedTermCount),
-			wantAntiAffinityPodsMap: topologyToMatchedTermCount{
-				{key: "hostname", value: "nodeA"}: 1,
-			},
-		},
-		{
-			name:         "incoming pod's anti-affinity has more than one term - existing pod violates partial term - case 2",
-			existingPods: []*v1.Pod{normalPodB},
-			nodes:        []*v1.Node{nodeA},
-			pod: &v1.Pod{
-				ObjectMeta: metav1.ObjectMeta{Name: "affi-antiaffi"},
-				Spec: v1.PodSpec{
-					Affinity: &v1.Affinity{
-						PodAffinity: &v1.PodAffinity{
-							RequiredDuringSchedulingIgnoredDuringExecution: newPodAffinityTerms("aaa", "bbb"),
-						},
-						PodAntiAffinity: &v1.PodAntiAffinity{
-							RequiredDuringSchedulingIgnoredDuringExecution: newPodAffinityTerms("aaa", "bbb"),
-						},
-					},
-				},
-			},
-			wantAffinityPodsMap: make(topologyToMatchedTermCount),
-			wantAntiAffinityPodsMap: topologyToMatchedTermCount{
-				{key: "hostname", value: "nodeA"}: 1,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			s := nodeinfosnapshot.NewSnapshot(nodeinfosnapshot.CreateNodeInfoMap(tt.existingPods, tt.nodes))
-			l, _ := s.NodeInfos().List()
-			gotAffinityPodsMap, gotAntiAffinityPodsMap, err := getTPMapMatchingIncomingAffinityAntiAffinity(tt.pod, l)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("getTPMapMatchingIncomingAffinityAntiAffinity() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if !reflect.DeepEqual(gotAffinityPodsMap, tt.wantAffinityPodsMap) {
-				t.Errorf("getTPMapMatchingIncomingAffinityAntiAffinity() gotAffinityPodsMap = %#v, want %#v", gotAffinityPodsMap, tt.wantAffinityPodsMap)
-			}
-			if !reflect.DeepEqual(gotAntiAffinityPodsMap, tt.wantAntiAffinityPodsMap) {
-				t.Errorf("getTPMapMatchingIncomingAffinityAntiAffinity() gotAntiAffinityPodsMap = %#v, want %#v", gotAntiAffinityPodsMap, tt.wantAntiAffinityPodsMap)
-			}
-		})
-	}
-}
-
 func TestGetTPMapMatchingSpreadConstraints(t *testing.T) {
 	fooSelector := st.MakeLabelSelector().Exists("foo").Obj()
 	barSelector := st.MakeLabelSelector().Exists("bar").Obj()
